// apps/humayro/lib/api/model/organizationAdvertisementRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */

export interface OrganizationAdvertisementRequest {
  organizationId: number;
  videoAttachmentId: number;
  /**
     * @minLength 0
     * @maxLength 150
     */
  title?: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  redirectUrl?: string;
  active?: boolean;
  sortOrder?: number;
  startsAt?: string;
  endsAt?: string;
}